pyproject.toml
[tool.poetry]
name = "kk-scheduler"
version = "0.1.0"
description = ""
authors = ["kkaneko <gorutiru@gmail.com>"]
readme = "README.md"
packages = [{include = "kk_scheduler"}]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2.2.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
